Prilagođeni fragmenti za korisničke postavke i stanje bodova te uklanjanje suvišnog koda na glavnoj aktivnosti.
